Convert to storing milliseconds since Unix epoch in DB

+++ blitz/data/models.py
@@ -3,7 +3,7 @@ __author__ = 'mecharius'
 import json
 
 from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy import Column, String, Integer, Boolean, ForeignKey, Numeric
+from sqlalchemy import Column, String, Integer, Boolean, ForeignKey
 from sqlalchemy.orm import relationship, backref
 
 # set up the base model
@@ -17,7 +17,7 @@ class Notification(SQL_BASE):
     __tablename__ = 'notifications'
 
     id = Column(Integer, primary_key=True)
-    timeLogged = Column(Numeric(12,3))
+    timeLogged = Column(Integer)
     severity = Column(Integer)
     description = Column(String)
 
@@ -27,7 +27,7 @@ class Notification(SQL_BASE):
         """
         return {
             "id": self.id,
-            "timeLogged": self.timeLogged,
+            "timeLogged": float(self.timeLogged),
             "severity": self.severity,
             "description": self.description
         }
@@ -44,7 +44,7 @@ class Reading(SQL_BASE):
 
     id = Column(Integer, primary_key=True)
     sessionId = Column(Integer)
-    timeLogged = Column(Numeric(12,3))
+    timeLogged = Column(Integer)
     categoryId = Column(Integer, ForeignKey('category.id'))
     value = Column(String)
 
@@ -57,7 +57,7 @@ class Reading(SQL_BASE):
         return {
             "id": self.id,
             "sessionId": self.sessionId,
-            "timeLogged": self.timeLogged,
+            "timeLogged": float(self.timeLogged),
             "categoryId": self.categoryId,
             "value": self.value
         }
@@ -73,8 +73,8 @@ class Session(SQL_BASE):
 
     id = Column(Integer, primary_key=True)
     available = Column(Boolean, default=False)
-    timeStarted = Column(Numeric(12,3))
-    timeStopped = Column(Numeric(12,3))
+    timeStarted = Column(Integer)
+    timeStopped = Column(Integer)
     numberOfReadings = Column(Integer)
 
     def to_dict(self):
@@ -84,8 +84,8 @@ class Session(SQL_BASE):
         return {
             "id": self.id,
             "available": self.available,
-            "timeStarted": self.timeStarted,
-            "timeStopped": self.timeStopped,
+            "timeStarted": float(self.timeStarted),
+            "timeStopped": float(self.timeStopped),
             "numberOfReadings": self.numberOfReadings
         }
 
@@ -147,7 +147,7 @@ class Cache(SQL_BASE):
     __tablename__ = 'cache'
 
     id = Column(Integer, primary_key=True)
-    timeLogged = Column(Numeric(12,3))
+    timeLogged = Column(Integer)
     categoryId = Column(Integer)
     value = Column(String)
 
@@ -157,7 +157,7 @@ class Cache(SQL_BASE):
         """
         return {
             "id": self.id,
-            "timeLogged": self.timeLogged,
+            "timeLogged": float(self.timeLogged),
             "categoryId": self.categoryId,
             "value": self.value
         }
